Rename contenttypes.py to object_types.py
